requirements/production.txt
-r base.txt

django-s3-folder-storage==0.2
